Crate narwhal_executor

Modules

__mock_MockExecutionState

__mock_MockExecutionState_ExecutionState

Macros

bail

ensure

Structs

ExecutionIndices

The state of the subscriber keeping track of the transactions that have already been executed. It ensures we do not process twice the same transaction despite crash-recovery.

Executor

A client subscribing to the consensus output and executing every transaction.

MockExecutionState

Enums

SubscriberError

Traits

ExecutionState

Functions

get_restored_consensus_output

Type Aliases

SerializedTransaction

Convenience type representing a serialized transaction.

SerializedTransactionDigest

Convenience type representing a serialized transaction digest.

SubscriberResult